+++ Sources/NablaScheduling/Views/Scenes/ScheduleAppointment/ScheduleAppointmentDelegate.swift
@@ -7,4 +7,6 @@ public protocol ScheduleAppointmentDelegate: AnyObject {
     /// Once the payment is done, and your backend did update the appointment, call the `completion` handler.
     /// You can return some `UINavigationController`, but let the SDK handle the presentation and dismissal of this controller.
     func paymentViewController(for appointment: Appointment, completion: @escaping (Result<Void, Error>) -> Void) -> UIViewController
+    
+    func scheduleAppointmentDidSucceed(with appointment: Appointment)
 }